refactor: extract IssueReproductionStep.swift from IssueCoreSubtypes.swift
Splits the `IssueReproductionStep` struct out of IssueCoreSubtypes.swift
so that file holds only the category / severity enums.

Byte-identical move of lines 21-52. Both files import Foundation.

IssueCoreSubtypes.swift: 147 → 114 lines. Tests: 667.
